fix(regression): display restore errors on UI
Fixes #506

When an error occurs on restore, we now display the error message in a flushbar & redirecting user to the same page, allowing users to retry manually when ready. Previously, we redirected user to the first page & didn't display any errors due to a regression while moving the business logic outside the UI.

The generic error message is now displayed in a WarningBox for consistency. We've also created EnterMnemonicsPageArguments to better structure the data passed to EnterMnemonicsPage.